// types/MentorType.ts
export interface MentorType {
  _id: string;
  name: string;
  email: string;
  phone: string;
  whatsapp: string;
  ipAddress: string;
  ipDetails: any;
  college: string;
  year: number;
  github: string;
  linkedIn: string;
  projectName: string;
  projectLink: string;
  projectTags: string[];
  projectDescription: string;
  question1: string;
  answer1: string;
  question2: string;
  answer2: string;
  isSelected: boolean;
  isBanned: boolean;
}